Update credit card migraiton apr to be decimal.

+++ db/migrate/20180418055620_create_credit_cards.rb
@@ -3,7 +3,7 @@ class CreateCreditCards < ActiveRecord::Migration[5.1]
     create_table :credit_cards do |t|
       t.string :name
       t.text :description
-      t.integer :apr
+      t.decimal :apr
       t.integer :limit
       t.integer :balance
       t.integer :accrued_interest
